docker/requirements.txt
nibabel==4.0.1
numpy==1.22.4
Pillow==9.2.0
segmentation_models_pytorch==0.2.1
SimpleITK==2.1.1.2
torchvision==0.12.0
